+++ myft/ui/myft-buttons/get-data-from-inputs.js
@@ -1,0 +1,16 @@
+export default function (inputs) {
+	const meta = {};
+
+	inputs.forEach((input) => {
+		if (input.name.startsWith('_rel.')) {
+			const key = input.name.slice('_rel.'.length);
+			meta._rel = meta._rel || {};
+			meta._rel[key] = input.value;
+
+		} else if (input.type === 'hidden') {
+			meta[input.name] = input.value;
+		}
+	});
+
+	return meta;
+}
